backend/pyproject.toml
[project]
name = "ai-qr-ordering-backend"
version = "0.1.0"
description = "AI QR Ordering System - Django Backend"
authors = [
    {name = "k.yook"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (==4.2.17)",
    "djangorestframework (==3.15.2)",
    "djangorestframework-simplejwt (==5.4.0)",
    "django-cors-headers (==4.6.0)",
    "channels (==4.2.0)",
    "channels-redis (==4.2.1)",
    "daphne (==4.1.2)",
    "agno (>=2.4.0)",
    "openai (>=1.66.0)",
    "anthropic (>=0.43.0)",
    "psycopg2-binary (==2.9.10)",
    "pillow (==11.1.0)",
    "python-decouple (==3.8)",
    "stripe (>=14.3.0,<15.0.0)",
    "google-auth (>=2.48.0,<3.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "cryptography (>=46.0.5,<47.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "8.3.4"
pytest-django = "4.9.0"
factory-boy = "3.3.1"
pytest-asyncio = "0.25.3"
ruff = ">=0.9.0"

